disk-mapper: allow building without cgo dependencies for linting
