+++ disk-mapper/internal/test/benchmark_test.go
@@ -1,4 +1,4 @@
-//go:build integration
+//go:build integration && cgo
 
 /*
 Copyright (c) Edgeless Systems GmbH
